+++ src/main/java/nova/command/AddEventCommand.java
@@ -1,3 +1,5 @@
+package nova.command;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
