+++ src/main/java/nova/command/DeleteCommand.java
@@ -1,3 +1,7 @@
+package nova.command;
+
+import nova.command.Command;
+
 public class DeleteCommand implements Command {
     private final int index;
