+++ src/main/java/nova/command/MarkCommand.java
@@ -1,15 +1,19 @@
-public class UnmarkCommand implements Command {
+package nova.command;
+
+import nova.command.Command;
+
+public class MarkCommand implements Command {
     private final int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, UiNova ui, Storage storage) throws NovaException {
         Task task = taskList.getTask(index);
-        task.markAsNotDone();
+        task.markAsDone();
         storage.saveTasks(taskList);
-        ui.printTaskUpdatedMessage("OK, I've marked this task as not done yet:", task);
+        ui.printTaskUpdatedMessage("Nice! I've marked this task as done:", task);
     }
 }
