+++ src/main/java/nova/command/UnmarkCommand.java
@@ -1,15 +1,19 @@
-public class MarkCommand implements Command {
+package nova.command;
+
+import nova.command.Command;
+
+public class UnmarkCommand implements Command {
     private final int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, UiNova ui, Storage storage) throws NovaException {
         Task task = taskList.getTask(index);
-        task.markAsDone();
+        task.markAsNotDone();
         storage.saveTasks(taskList);
-        ui.printTaskUpdatedMessage("Nice! I've marked this task as done:", task);
+        ui.printTaskUpdatedMessage("OK, I've marked this task as not done yet:", task);
     }
 }
