+++ src/main/java/nova/exception/NovaException.java
@@ -1,3 +1,5 @@
+package nova.exception;
+
 public class NovaException extends RuntimeException {
     public NovaException(String message) {
         super(message);
