+++ src/main/java/nova/task/Deadline.java
@@ -1,7 +1,9 @@
+package nova.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
     protected LocalDateTime by;
     private static final DateTimeFormatter SAVE_FORMATTER = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
