+++ src/main/java/nova/task/Event.java
@@ -1,3 +1,5 @@
+package nova.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
