+++ src/main/java/nova/task/Todo.java
@@ -1,3 +1,5 @@
+package nova.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
