+++ src/main/java/nova/ui/Nova.java
@@ -1,4 +1,8 @@
-import java.util.Scanner;
+package nova.ui;
+
+import nova.command.Command;
+import nova.exception.NovaException;
+import nova.task.TaskList;
 
 public class Nova {
     private Storage storage;
