+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Nova
+ Hello! I'm nova.ui.Nova
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
